chore: reorganize code into a workspace
