+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "spark"]
-	path = spark
+[submodule "core/spark"]
+	path = core/spark
 	url = https://github.com/apache/spark
