+++ core/src/catalog.rs
