+++ core/src/client/mod.rs
@@ -1,3 +1,5 @@
+//! Generic implementation of ChannelBuilder and SparkConnectClient
+
 use std::collections::HashMap;
 use std::env;
 use std::str::FromStr;
