+++ core/src/functions/mod.rs
@@ -5,12 +5,22 @@ use crate::spark;
 use crate::DataFrame;
 
 use crate::column::Column;
-use expressions::{ToExpr, ToLiteralExpr};
-
-use crate::generate_functions;
-use crate::utils::invoke_func;
+use expressions::{ToExpr, ToLiteralExpr, ToVecExpr};
 
 use rand::random;
+
+pub fn invoke_func<T: ToVecExpr>(name: &str, args: T) -> Column {
+    Column::from(spark::Expression {
+        expr_type: Some(spark::expression::ExprType::UnresolvedFunction(
+            spark::expression::UnresolvedFunction {
+                function_name: name.to_string(),
+                arguments: args.to_vec_expr(),
+                is_distinct: false,
+                is_user_defined_function: false,
+            },
+        )),
+    })
+}
 
 /// Create a column from a &str
 pub fn col(value: &str) -> Column {
@@ -258,6 +268,45 @@ pub fn desc_nulls_first<T: ToLiteralExpr>(col: T) -> Column {
 
 pub fn desc_nulls_last<T: ToLiteralExpr>(col: T) -> Column {
     Column::from(col.to_literal_expr()).desc_nulls_last()
+}
+
+macro_rules! generate_functions {
+    (no_args: $($func_name:ident),*) => {
+        $(
+            pub fn $func_name() -> Column {
+                let empty_args: Vec<Column> = vec![];
+                invoke_func(stringify!($func_name), empty_args)
+            }
+        )*
+    };
+    (one_col: $($func_name:ident),*) => {
+        $(
+            pub fn $func_name<T: expressions::ToExpr>(col: T) -> Column
+            where
+                Vec<T>: expressions::ToVecExpr,
+            {
+                invoke_func(stringify!($func_name), vec![col])
+            }
+        )*
+    };
+    (two_cols: $($func_name:ident),*) => {
+        $(
+            pub fn $func_name<T: expressions::ToExpr>(col1: T, col2: T) -> Column
+            where
+                Vec<T>: expressions::ToVecExpr,
+            {
+                invoke_func(stringify!($func_name), vec![col1, col2])
+            }
+        )*
+    };
+    (multiple_cols: $($func_name:ident),*) => {
+        $(
+            pub fn $func_name<T: expressions::ToVecExpr>(cols: T) -> Column
+            {
+                invoke_func(stringify!($func_name), cols)
+            }
+        )*
+    };
 }
 
 // functions that require no arguments
