+++ core/src/streaming/mod.rs
