+++ rust/examples/delta.rs
