+++ rust/examples/reader.rs
