+++ rust/examples/readstream.rs
