+++ rust/examples/writer.rs
