+++ rust/src/lib.rs
@@ -1,0 +1,4 @@
+/**
+ * The spark-connect-rs crate is currently just a meta-package shim for spark-connect-core
+ */
+pub use spark_connect_core::*;
